requirements.txt
# Dev environment
pip
autopep8

# App
flask
requests
Pillow
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
# twilio (removed)